Accept friendship completed
