+++ app.js
@@ -120,8 +120,9 @@ app.use("/login", anonRouter);				// Sets the access to the login (only anonymou
 app.use("/signup", anonRouter);				// Sets the access to the register (only anonymous)
 app.use('/api/friends', userTokenRouter);	// Sets the api access
 app.use('/api/message', userTokenRouter);
-app.use("/users", authRouter);				// Sets the access to the rest of the app (only auth)
-app.use("/friend/*", authRouter);
+app.use("/users*", authRouter);				// Sets the access to the rest of the app (only auth)
+app.use("/requests*", authRouter);
+app.use("/friends*", authRouter);
 app.use(express.static("public"));			// Sets the static folder
 
 // Headers
@@ -187,7 +188,6 @@ app.encrypt = (string) => {
 \*****************************************************************************/
 
 require("./routes/rusers")(app, dbManager);									// Users controller
-require("./routes/rrequests")(app, dbManager);								// Requests controller
 require("./routes/rfriends")(app, dbManager);								// Friends controller
 require("./routes/rapisocial")(app, dbManager);								// API controller
