+++ views/base.html
@@ -21,7 +21,7 @@
         <span class="icon-bar"></span>
         <span class="icon-bar"></span>                        
       </button>
-      <img src="/img/logo.svg" alt="logo" height="50px" style="margin-right: 15px; color: #9d9d9d"/>
+      <a href="/"><img src="/img/logo.svg" alt="logo" height="50px" style="margin-right: 15px"/></a>
     </div>
     <div class="collapse navbar-collapse" id="myNavbar">
       <ul class="nav navbar-nav">
