+++ views/requests.html
@@ -25,7 +25,7 @@
                     <td></td>
                     <td>
                         <!-- Accept button -->
-                        <a type="button" class="btn btn-info" href="#">
+                        <a type="button" class="btn btn-info" href="/friends/accept/{{f.email}}">
                             <span class="glyphicon glyphicon-ok"></span> Accept
                         </a>
                     </td>
